### 2023-09-12 #5840 @gkthiruvathukal
Thanks for submitting your paper to JOSS @jjmcintosh. Currently, there isn't a JOSS editor assigned to your paper.  We need one cognizant editor and two reviewers to proceed.

Please note that this track is facing a _backlog_ of submissions. Please take a look at the topic editors from our [editorial board](https://joss.theoj.org/about#editorial_board) and let me know up to three editorial board members whose interests overlap with your submission. Please list their GitHub IDs but do not use an at (@) symbol in front of their GitHub username.

Please respond within a week of this comment. If I do not receive a response, I will seek input from our editorial board to determine whether this submission is within scope.
