### 2023-12-02 #5840 @osorensen
👋 @mdbloice, @maieroli2010, @kuadrat, would any of you be willing to review this submission for JOSS? It seems to be directly within your areas of expertise.
