### 2023-12-04 #6120 @editorialbot
[REVIEW]: DICaugment: A Python Package for 3D Medical Imaging Augmentation
**Submitting author:** @jjmcintosh (<a href="http://orcid.org/0009-0008-2573-180X">Jacob McIntosh</a>)
**Repository:** https://github.com/DIDSR/DICaugment
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.7
