### 2023-12-05 #6120 @MaierOli2010
Hi all,

thank you for this nice work.
I've just started to look through it and have seen the following issues:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I did not find any automated testing nor instructions on how to perform testing of the software. There is, however, a tests subfolder so I assume some testing is done in one way or another. Please either state how to perform the testing manually or, ideally, use some automated software to perform testing (like Jenkins or Travis).

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

There are no guidelines on how to report problems/bugs or how to contribute. Please include an appropriate document such as  from this template: https://gist.github.com/briandk/3d2e8b3ec8daf5a27a62 and adapt it to your needs.

While not mandatory, a code of conduct would also be recommended.

>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I've also noticed that not all parameters of the API are documented in each function. E.g. always_apply is not documented in all functions at https://dicaugment.readthedocs.io/en/latest/dicaugment.augmentations.html. Please make sure that all parameters of the public API are documented accordingly.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

I do miss a broader overview of existing software for data augmentation. Albumentations is not the only package out there to do image augmentation. Some examples that I quickly found are TorchIO, MONAI, DLTK (also offers some augmentation techniques), MedicalTorch also claims to have some elastic transformations available to augment data. Finally, there is a tutorial from SimpleITK to apply data augmentation for 2D and 3D images available at https://simpleitk.org/SPIE2019_COURSE/03_data_augmentation.html. Please clearly state how you differentiate to these other tools, especially the claim of 3D augmentation seems to be covered at least by SimpleITK.

To this end, I also think that the list of references is not complete as these packages, and the corresponding publications if applicable, are certainly missing.

> Contribution and authorship: Has the submitting author (@jjmcintosh) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

While I can clearly see the contribution of the main author and also the last author as mentioned in the readme, the others are not to be found in any commit and are also not mentioned in your own readme. Can you maybe clarify the contributions? I do not doubt the contributions but it is hard to follow if there are no commits and even your own readme does not mention them as contributors.

While also not mandatory, a live example of the software using e.g. Google Colab could greatly enhance the user experience. The provided guides seem sufficient to understand the functionality.

Testing the basic example in the readme gives an error:
```
import dicaugment as dca
# Define the augmentation pipeline
transform = dca.Compose([
    dca.Rotate(p=0.5, limit=20, interpolation=1),
    dca.RandomCrop(p=0.5, size=(64, 64, 64))
])
```
> TypeError: RandomCrop.__init__() got an unexpected keyword argument 'size'

This stems from the fact that the transform expects height, width, and depth in its definition.
Please make sure that all examples work as given in the readme or on RTD.

Overall I think the toolbox does offer unique functionality in the context of CT imaging using the NPS for noise augmentation.

Cheers,
Oliver
