### 2023-12-14 db74ba3
chore: add automated testing and linter
