+++ tests/test_keypoint.py
@@ -29,7 +29,11 @@ def test_convert_keypoint_to_dicaugment(kp, source_format, expected):
     image = np.ones((100, 100, 100))
 
     converted_keypoint = convert_keypoint_to_dicaugment(
-        kp, rows=image.shape[0], cols=image.shape[1], slices=image.shape[2], source_format=source_format
+        kp,
+        rows=image.shape[0],
+        cols=image.shape[1],
+        slices=image.shape[2],
+        source_format=source_format,
     )
     assert converted_keypoint == expected
 
@@ -47,7 +51,11 @@ def test_convert_keypoint_to_dicaugment(kp, source_format, expected):
 def test_convert_keypoint_from_dicaugment(kp, target_format, expected):
     image = np.ones((100, 100, 100))
     converted_keypoint = convert_keypoint_from_dicaugment(
-        kp, rows=image.shape[0], cols=image.shape[1], slices=image.shape[2], target_format=target_format
+        kp,
+        rows=image.shape[0],
+        cols=image.shape[1],
+        slices=image.shape[2],
+        target_format=target_format,
     )
     assert converted_keypoint == expected
 
@@ -64,10 +72,18 @@ def test_convert_keypoint_from_dicaugment(kp, target_format, expected):
 def test_convert_keypoint_to_dicaugment_and_back(kp, keypoint_format):
     image = np.ones((100, 100, 100))
     converted_kp = convert_keypoint_to_dicaugment(
-        kp, rows=image.shape[0], cols=image.shape[1], slices=image.shape[2], source_format=keypoint_format
+        kp,
+        rows=image.shape[0],
+        cols=image.shape[1],
+        slices=image.shape[2],
+        source_format=keypoint_format,
     )
     converted_back_kp = convert_keypoint_from_dicaugment(
-        converted_kp, rows=image.shape[0], cols=image.shape[1], slices=image.shape[2], target_format=keypoint_format
+        converted_kp,
+        rows=image.shape[0],
+        cols=image.shape[1],
+        slices=image.shape[2],
+        target_format=keypoint_format,
     )
     assert converted_back_kp == kp
 
@@ -76,13 +92,25 @@ def test_convert_keypoints_to_dicaugment():
     keypoints = [(20, 30, 40, 50, 60), (30, 40, 50, 60, 70, 99)]
     image = np.ones((100, 100, 100))
     converted_keypoints = convert_keypoints_to_dicaugment(
-        keypoints, rows=image.shape[0], cols=image.shape[1], slices=image.shape[2], source_format="xyzas"
+        keypoints,
+        rows=image.shape[0],
+        cols=image.shape[1],
+        slices=image.shape[2],
+        source_format="xyzas",
     )
     converted_keypoint_1 = convert_keypoint_to_dicaugment(
-        keypoints[0], rows=image.shape[0], cols=image.shape[1], slices=image.shape[2], source_format="xyzas"
+        keypoints[0],
+        rows=image.shape[0],
+        cols=image.shape[1],
+        slices=image.shape[2],
+        source_format="xyzas",
     )
     converted_keypoint_2 = convert_keypoint_to_dicaugment(
-        keypoints[1], rows=image.shape[0], cols=image.shape[1], slices=image.shape[2], source_format="xyzas"
+        keypoints[1],
+        rows=image.shape[0],
[…]
