+++ dicaugment/augmentations/dropout/__init__.py
@@ -1,4 +1,2 @@
-#from .channel_dropout import *
 from .coarse_dropout import *
 from .grid_dropout import *
-#from .mask_dropout import *
