+++ tools/make_kp_gif.py
@@ -8,7 +8,7 @@ import sys
 
 # SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
 # sys.path.append(os.path.dirname(SCRIPT_DIR))
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
 import dicaugment as A
 import random
@@ -16,6 +16,7 @@ from PIL import Image
 import glob
 
 import shutil
+
 
 def get_cube_edge_pairs(bbox):
     min_x, min_y, min_z, max_x, max_y, max_z = bbox
@@ -33,21 +34,31 @@ def get_cube_edge_pairs(bbox):
         [(max_x, min_y, min_z), (max_x, min_y, max_z)],
         [(max_x, min_y, min_z), (max_x, max_y, min_z)],
         [(max_x, min_y, max_z), (max_x, max_y, max_z)],
-        [(max_x, max_y, min_z), (max_x, max_y, max_z)]
+        [(max_x, max_y, min_z), (max_x, max_y, max_z)],
     ]
 
     return edge_pairs
 
 
-def gif_ify(directory, ext = ".png"):
-        frames = [Image.open(im) for im in sorted(glob.glob("{}*{}".format(directory, ext)))]
-        frames_rev = [Image.open(im) for im in sorted(glob.glob("{}*{}".format(directory, ext)))][1:-1][::-1]
-        first_frame = frames[0]
-        first_frame.save("kp_example.gif", append_images = frames[1:] + frames_rev, optimize= False, save_all= True, duration = 100, loop = 0)
+def gif_ify(directory, ext=".png"):
+    frames = [
+        Image.open(im) for im in sorted(glob.glob("{}*{}".format(directory, ext)))
+    ]
+    frames_rev = [
+        Image.open(im) for im in sorted(glob.glob("{}*{}".format(directory, ext)))
+    ][1:-1][::-1]
+    first_frame = frames[0]
+    first_frame.save(
+        "kp_example.gif",
+        append_images=frames[1:] + frames_rev,
+        optimize=False,
+        save_all=True,
+        duration=100,
+        loop=0,
+    )
 
 
 def main():
-
     os.chdir("./tools")
 
     if os.path.exists("./out"):
@@ -57,55 +68,62 @@ def main():
 
     image = np.zeros((100, 100, 100, 3), dtype=np.uint8)
 
-    for i in range(10,40):
-        image[i:80-i, 80-i -2 :80-i, 20 + i] = [255, 0, 255]
-        image[i:80-i, i:i+2, 20 + i] = [255, 255, 0]
+    for i in range(10, 40):
+        image[i : 80 - i, 80 - i - 2 : 80 - i, 20 + i] = [255, 0, 255]
+        image[i : 80 - i, i : i + 2, 20 + i] = [255, 255, 0]
 
-        image[80-i -2 :80-i, i:80-i, 20 + i] = [0, 0, 255]
-        image[i:i+2, i:80-i, 20 + i] = [0, 255, 0]
-
-
+        image[80 - i - 2 : 80 - i, i : 80 - i, 20 + i] = [0, 0, 255]
+        image[i : i + 2, i : 80 - i, 20 + i] = [0, 255, 0]
 
     bboxes = [[9, 9, 29, 71, 71, 61]]
-    keypoints = [[10,10,30], [70,70,30], [10,70,30], [70,10,30], [40,40,59]]
-
+    keypoints = [[10, 10, 30], [70, 70, 30], [10, 70, 30], [70, 10, 30], [40, 40, 59]]
 
     random.seed(42)
[…]
