### 2023-12-15 8f1aaad
chore: add automated testing and linter (#2)
* chore: add automated testing and linter

* fix: add pytest-cov

---------

Co-authored-by: jjmcintosh <jjmcintosh@github.com>
