+++ tools/make_transforms_docs.py
@@ -26,8 +26,11 @@ def make_augmentation_docs_link(cls):
     return (
         "[{cls.__name__}](https://dicaugment.readthedocs.io/en/latest/{module_page}.html#{cls.__module__}.{cls.__name__})"
     ).format(module_page=module_page, cls=cls)
-#dicaugment.readthedocs.io/en/latest/dicaugment.augmentations.crops.html#dicaugment.augmentations.crops.transforms.CenterCrop
-#https://dicaugment.readthedocs.io/en/latest/dicaugment.augmentations.dropout.html#dicaugment.augmentations.dropout.coarse_dropout.CoarseDropout
+
+
+# dicaugment.readthedocs.io/en/latest/dicaugment.augmentations.crops.html#dicaugment.augmentations.crops.transforms.CenterCrop
+# https://dicaugment.readthedocs.io/en/latest/dicaugment.augmentations.dropout.html#dicaugment.augmentations.dropout.coarse_dropout.CoarseDropout
+
 
 class Targets(Enum):
     IMAGE = "Image"
@@ -41,7 +44,9 @@ def parse_args():
     subparsers = parser.add_subparsers(help="Commands", dest="command")
     subparsers.add_parser("make")
     check_parser = subparsers.add_parser("check")
-    check_parser.add_argument("filepath", type=str, help="Path to a file that should be checked")
+    check_parser.add_argument(
+        "filepath", type=str, help="Path to a file that should be checked"
+    )
     return parser.parse_args()
 
 
@@ -55,8 +60,14 @@ def get_transforms_info():
     transforms_info = {}
     members = inspect.getmembers(dicaugment)
     for name, cls in members:
-        if inspect.isclass(cls) and issubclass(cls, dicaugment.BasicTransform) and name not in IGNORED_CLASSES:
-            if "DeprecationWarning" in inspect.getsource(cls) or "FutureWarning" in inspect.getsource(cls):
+        if (
+            inspect.isclass(cls)
+            and issubclass(cls, dicaugment.BasicTransform)
+            and name not in IGNORED_CLASSES
+        ):
+            if "DeprecationWarning" in inspect.getsource(
+                cls
+            ) or "FutureWarning" in inspect.getsource(cls):
                 continue
 
             targets = {Targets.IMAGE}
@@ -64,7 +75,8 @@ def get_transforms_info():
                 targets.add(Targets.MASKS)
 
             if (
-                hasattr(cls, "apply_to_bbox") and cls.apply_to_bbox is not dicaugment.DualTransform.apply_to_bbox
+                hasattr(cls, "apply_to_bbox")
+                and cls.apply_to_bbox is not dicaugment.DualTransform.apply_to_bbox
             ) or (
                 hasattr(cls, "apply_to_bboxes")
                 and cls.apply_to_bboxes is not dicaugment.DualTransform.apply_to_bboxes
@@ -73,10 +85,12 @@ def get_transforms_info():
 
             if (
                 hasattr(cls, "apply_to_keypoint")
-                and cls.apply_to_keypoint is not dicaugment.DualTransform.apply_to_keypoint
+                and cls.apply_to_keypoint
+                is not dicaugment.DualTransform.apply_to_keypoint
             ) or (
                 hasattr(cls, "apply_to_keypoints")
-                and cls.apply_to_keypoints is not dicaugment.DualTransform.apply_to_keypoints
+                and cls.apply_to_keypoints
+                is not dicaugment.DualTransform.apply_to_keypoints
             ):
                 targets.add(Targets.KEYPOINTS)
 
@@ -109,16 +123,19 @@ def make_transforms_targets_table(transforms_info, header):
     column_widths = [max(len(r) for r in column) for column in zip(*rows)]
     lines = [
[…]
