+++ dicaugment/augmentations/geometric/resize.py
@@ -11,7 +11,7 @@ from ...core.transforms_interface import (
     DualTransform,
     KeypointInternalType,
     to_tuple,
-    INTER_LINEAR
+    INTER_LINEAR,
 )
 from . import functional as F
 from ..dicom import functional as Fdicom
@@ -37,7 +37,13 @@ class RandomScale(DualTransform):
         uint8, uint16, int16, float32
     """
 
-    def __init__(self, scale_limit: float = 0.1, interpolation: int = INTER_LINEAR, always_apply=False, p=0.5):
+    def __init__(
+        self,
+        scale_limit: float = 0.1,
+        interpolation: int = INTER_LINEAR,
+        always_apply=False,
+        p=0.5,
+    ):
         super(RandomScale, self).__init__(always_apply, p)
         self.scale_limit = to_tuple(scale_limit, bias=1.0)
         self.interpolation = interpolation
@@ -54,12 +60,15 @@ class RandomScale(DualTransform):
 
     def apply_to_keypoint(self, keypoint, scale=1, **params):
         return F.keypoint_scale(keypoint, scale, scale, scale)
-    
-    def apply_to_dicom(self, dicom: DicomType, scale = 1, **params) -> DicomType:
+
+    def apply_to_dicom(self, dicom: DicomType, scale=1, **params) -> DicomType:
         return Fdicom.dicom_scale(dicom, scale, scale)
 
     def get_transform_init_args(self):
-        return {"interpolation": self.interpolation, "scale_limit": to_tuple(self.scale_limit, bias=-1.0)}
+        return {
+            "interpolation": self.interpolation,
+            "scale_limit": to_tuple(self.scale_limit, bias=-1.0),
+        }
 
 
 class LongestMaxSize(DualTransform):
@@ -90,7 +99,11 @@ class LongestMaxSize(DualTransform):
         self.max_size = max_size
 
     def apply(
-        self, img: np.ndarray, max_size: int = 1024, interpolation: int = INTER_LINEAR, **params
+        self,
+        img: np.ndarray,
+        max_size: int = 1024,
+        interpolation: int = INTER_LINEAR,
+        **params
     ) -> np.ndarray:
         return F.longest_max_size(img, max_size=max_size, interpolation=interpolation)
 
@@ -98,15 +111,19 @@ class LongestMaxSize(DualTransform):
         # Bounding box coordinates are scale invariant
         return bbox
 
-    def apply_to_keypoint(self, keypoint: KeypointInternalType, max_size: int = 1024, **params) -> KeypointInternalType:
+    def apply_to_keypoint(
+        self, keypoint: KeypointInternalType, max_size: int = 1024, **params
+    ) -> KeypointInternalType:
         height = params["rows"]
         width = params["cols"]
         depth = params["slices"]
 
         scale = max_size / max([height, width, depth])
         return F.keypoint_scale(keypoint, scale, scale, scale)
-    
-    def apply_to_dicom(self, dicom: DicomType, max_size: int = 1024, **params) -> DicomType:
+
+    def apply_to_dicom(
+        self, dicom: DicomType, max_size: int = 1024, **params
+    ) -> DicomType:
         height = params["rows"]
         width = params["cols"]
         depth = params["slices"]
@@ -114,7 +131,11 @@ class LongestMaxSize(DualTransform):
         return Fdicom.dicom_scale(dicom, scale, scale)
 
     def get_params(self) -> Dict[str, int]:
-        return {"max_size": self.max_size if isinstance(self.max_size, int) else random.choice(self.max_size)}
+        return {
+            "max_size": self.max_size
+            if isinstance(self.max_size, int)
+            else random.choice(self.max_size)
+        }
 
     def get_transform_init_args_names(self) -> Tuple[str, ...]:
[…]
