+++ dicaugment/__init__.py
@@ -10,14 +10,14 @@ from .core.bbox_utils import BboxParams
 from .core.keypoints_utils import KeypointParams
 
 try:
-    from .pytorch.transforms import * 
+    from .pytorch.transforms import *
 except ImportError:
     # torch is not installed by default, so we import stubs.
     # Run `pip install -U albumentations[torch] if you need augmentations from torch.`
     from .pytorch.stubs import *  # type: ignore
 
 try:
-    from .tensorflow.transforms import * 
+    from .tensorflow.transforms import *
 except ImportError:
     # tensorflow is not installed by default, so we import stubs.
     # Run `pip install -U albumentations[tensorflow] if you need augmentations from tensorflow.`

[…]
