### 2023-12-15 1e85993
docs(docstring): add always_apply to docstrings
