### 2023-12-18 58094d3
chore(template): update typo in pull request template
