+++ .github/pull_request_template.md
@@ -1,0 +1,13 @@
+### Title
+
+
+### Description
+
+
+### Supporting Evidence
+
+
+### Checklist
+- [ ] The commit messages adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and follow the syntax of [github-tag-action](https://github.com/anothrNick/github-tag-action)
+- [ ] Tests for the changes have been added (for bug fixes / features)
+- [ ] Docs have been added/updated (for bug fixes/features)
