+++ .github/workflows/auto_tag.yml
@@ -4,7 +4,7 @@ on:
     types:
       - closed
     branches:
-      - master
+      - main
 
 jobs:
   build:
@@ -24,3 +24,4 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         WITH_V: true
         PRERELEASE: true
+        DEFAULT_BUMP: patch

[…]
