### 2024-01-03 cfadc8d
docs(paper): Update paper
