### 2024-01-04 #6120 @kuadrat
Congratulations to this great work, @jjmcintosh!

I believe that this project and paper can certainly be published in JOSS. However, there are a few points that first need to be addressed in order for me to be able to check all the boxes.

# Critical points

- [ ] **Contribution and authorship** @jjmcintosh's contributions are evident from the commit logs. Mehdi Farhangi is listed as an author in the README, though it is not evident from the repository nor paper, in which way both authors contributed. The same goes for the other authors listed in the paper. Please specify this, ideally in the README or a CONTRIBUTORS file _and_ in the paper.
- [ ] **Automated tests** Running the tests after cloning the repository gave `206 failed, 4611 passed`. Ensure that all tests pass in the version that is to be published. Also, the source code of the tests contains large numbers of commented out blocks of code, which indicates that some work on that front might still be required.
- [ ] **Documentation** The tutorial part of the docs is clearly written, easy to follow and enhanced by well-made, relevant images. However, the API documentation contains a lot of undocumented functions and class method. Also, important methods such as the `Composition` class' `__call__` method should be documented and it should be clear from the API doc alone - without the tutorial - that *this* is the intended way to make use of the class (if my understanding is correct here).
I can see that most Transform classes have methods of identical names, such as `apply`, which always serve the same purpose. Maybe [this stackoverflow answer](https://stackoverflow.com/a/65460426/5264510)) provides some possibilities for re-using documentation.
- [ ] **State of the field** I agree with the point brought up by @MaierOli2010 regarding the need to specify how `DICaugment` distinguishes itself from other 3D image augmentation software. I also found an additional package that seems to be quite similar: [volumentations](https://github.com/ZFTurbo/volumentations).
- [ ] **Typos**
    - [this title](https://dicaugment.readthedocs.io/en/latest/bbox_augmentation.html#annotatation-formats) is misspelled
    - there's probably a newline missing in the docstring to [BBoxSafeRandomCrop](https://dicaugment.readthedocs.io/en/latest/augmentations.crops.html#dicaugment.augmentations.crops.transforms.BBoxSafeRandomCrop)
    - paper line 65: the units "mm" need not be italic
    - paper, caption of Fig. 1, second to last line: Siemens is misspelled

# Additional points

The following points are not considered to be blocking publication in JOSS. They should be seen as suggestions to further improve the software package and facilitate future development and maintenance.

- [ ] **Code coverage** In order to assess whether the designed tests actually do a good job at testing the functionality of the code, I recommend the use of a test-coverage estimator, such as [codecov](https://about.codecov.io/language/python/)
- [ ] **Tests** I appreciate the fact that a large number of test cases for many different parameter combinations is considered. However, this results in a rather calculation intensive test suite. In development, it can be helpful to run tests often to immediately notice any potential problems that have been introduced. It might therefore be worth it to set up a separate set of tests that is less thorough and can be used regularly during development, e.g. on each commit or push. The complete list of tests could then be run on a less regular basis, e.g. when merging to main or before publishing a new version.
- [ ] **Provide example file** In order to follow along with some of the tutorials (e.g. [mask augmentation](https://dicaugment.readthedocs.io/en/latest/mask_augmentation.html)), the user is expected to have certain files (like a CT scan file with a DICOM header). While this can reasonably be expected from the target audience, it would still be helpful to provide users with an example file for which you ensure that the tutorial example works. Otherwise, in case of trouble, users will have to figure out if the problem lies in their file or their installation.
