+++ paper/paper.bib
@@ -129,3 +129,23 @@
   copyright = "http://onlinelibrary.wiley.com/termsAndConditions",
   language  = "en"
 }
+
+@article{Yaniv2017, 
+  title={SIMPLEITK image-analysis notebooks: A collaborative environment for Education and reproducible research},
+  volume={31},
+  DOI={10.1007/s10278-017-0037-8},
+  number={3},
+  journal={Journal of Digital Imaging},
+  author={Yaniv, Ziv and Lowekamp, Bradley C. and Johnson, Hans J. and Beare, Richard},
+  year={2017},
+  pages={290-303}
+}
+
+@article{Pérez-García2021,
+  title={Torchio: A python library for efficient loading, preprocessing, augmentation and patch-based sampling of medical images in deep learning},
+  volume={208}, DOI={10.1016/j.cmpb.2021.106236},
+  journal={Computer Methods and Programs in Biomedicine},
+  author={Pérez-García, Fernando and Sparks, Rachel and Ourselin, Sébastien},
+  year={2021},
+  pages={106236}
+} 
