+++ .github/workflows/auto_tag.yml
@@ -23,5 +23,5 @@ jobs:
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         WITH_V: true
-        PRERELEASE: true
+        PRERELEASE: false
         DEFAULT_BUMP: patch
