### 2024-01-05 #6120 @jjmcintosh
@kuadrat Thanks! I will get these points addressed soon.

> Automated tests Running the tests after cloning the repository gave 206 failed, 4611 passed. Ensure that all tests pass in the version that is to be published. Also, the source code of the tests contains large numbers of commented out blocks of code, which indicates that some work on that front might still be required.

I have a follow-up question here. What system are you using and what is the version of Python? I have recently added a GitHub Action for automated testing and there aren't any failed tests there so I'd like to get to the bottom of where there might be a discrepancy in the outcomes of the tests.
