### 2024-01-08 DIDSR/DICaugment#5 @kuadrat
Some failing tests 
This issue is part of https://github.com/openjournals/joss-reviews/issues/6120 and was opened in response to https://github.com/openjournals/joss-reviews/issues/6120#issuecomment-1879332136.

# Summary

Running `pytest` gives `206 failed, 4611 passed`.

# Background

## System information

OS: Linux Debian 12 (bookworm)
Kernel: 6.1.0-16-amd64
Python version: 3.7.5
pyenv version: 2.3.5-2-g03a5d653
DICaugment version: github main branch, commit b752aad04e1cd2c3d8b67d

## Steps to reproduce

1. Create new virtual environment with `pyenv`, activate it locally:
```bash
$ pyenv virtualenv 3.7.5 dicaugment
$ pyenv local dicaugment
```

2. Update pip and install DICaugment and pytest:
```bash
$ pip install --upgrade pip
# upgrades successfully
$ pip install dicaugment
# installs successfully
$ pip install pytest
# installs successfully
```

3. Clone repo and move to DICaugment directory:
```bash
$ git clone git@github.com:DIDSR/DICaugment.git
$ cd DICaugment
```

4. Run pytest
```bash
$ pytest
# See attached file for output
```

## Output

[pytest-output.txt](https://github.com/DIDSR/DICaugment/files/13857880/pytest-output.txt)
