+++ tests/test_pytorch.py
@@ -112,18 +112,6 @@ def test_torch_to_tensor_v2_on_gray_scale_images():
     assert data["image"].dtype == torch.uint8
 
 
-# def test_torch_to_tensor_raises_runtime_error():
-#     with pytest.raises(RuntimeError) as exc_info:
-#         aug = ToTensor()  # noqa F841
-#     message = (
-#         "`ToTensor` is obsolete and it was removed from Albumentations. Please use `ToPytorch` instead - "
-#         "https://albumentations.ai/docs/api_reference/pytorch/transforms/"
-#         "#albumentations.pytorch.transforms.ToPytorch. "
-#         "\n\nIf you need `ToTensor` downgrade Albumentations to version 0.5.2."
-#     )
-#     assert str(exc_info.value) == message
-
-
 def test_with_replaycompose():
     aug = A.ReplayCompose([A.ToPytorch()])
     kwargs = {
@@ -142,49 +130,6 @@ def test_with_replaycompose():
     assert res["mask"].dtype == torch.uint8
     assert res2["image"].dtype == torch.uint8
     assert res2["mask"].dtype == torch.uint8
-
-
-# @pytest.mark.parametrize(
-#     ["brightness", "contrast", "saturation", "hue"],
-#     [
-#         [1, 1, 1, 0],
-#         [0.123, 1, 1, 0],
-#         [1.321, 1, 1, 0],
-#         [1, 0.234, 1, 0],
-#         [1, 1.432, 1, 0],
-#         [1, 1, 0.345, 0],
-#         [1, 1, 1.543, 0],
-#     ],
-# )
-# def test_brightness_contrast(brightness, contrast, saturation, hue):
-#     np.random.seed(0)
-#     img = np.random.randint(0, 256, [100, 100, 50], dtype=np.uint8)
-#     pil_image = Image.fromarray(img)
-
-#     transform = A.Compose(
-#         [
-#             A.RandomBrightnessContrast(
-#                 brightness=[brightness, brightness],
-#                 contrast=[contrast, contrast],
-#                 saturation=[saturation, saturation],
-#                 hue=[hue, hue],
-#                 p=1,
-#             )
-#         ]
-#     )
-
-#     pil_transform = ColorJitter(
-#         brightness=[brightness, brightness],
-#         contrast=[contrast, contrast],
-#         saturation=[saturation, saturation],
-#         hue=[hue, hue],
-#     )
-
-#     res1 = transform(image=img)["image"]
-#     res2 = np.array(pil_transform(pil_image))
-
-#     _max = np.abs(res1.astype(np.int16) - res2.astype(np.int16)).max()
-#     assert _max <= 2, "Max: {}".format(_max)
 
 
 def test_post_data_check():
