+++ .github/workflows/auto_test.yml
@@ -2,7 +2,6 @@ name: automated testing
 
 on:
   pull_request:
-    branches: [ "main" ]
 
 permissions:
   contents: read
