+++ .github/workflows/auto_test.yml
@@ -2,6 +2,7 @@ name: automated testing
 
 on:
   pull_request:
+    branches: [ "dev" ]
 
 permissions:
   contents: read
