### 2024-01-15 18569e4
build: fix workflow syntax
