+++ .github/workflows/auto_test.yml
@@ -2,7 +2,7 @@ name: automated testing
 
 on:
   pull_request:
-    branches: [ "dev" ]
+    branches: [ "main" ]
 
 permissions:
   contents: read
@@ -14,7 +14,7 @@ jobs:
 
     steps:
     - name: Check out code
-    - uses: actions/checkout@v3
+      uses: actions/checkout@v3
 
     - name: Set up Python 3.10
       uses: actions/setup-python@v3
