+++ .github/workflows/publish.yml
@@ -24,7 +24,7 @@ jobs:
 
     steps:
     - name: Check out Repository
-    - uses: actions/checkout@v3
+      uses: actions/checkout@v3
 
     - name: Set up Python
       uses: actions/setup-python@v3
@@ -59,7 +59,8 @@ jobs:
         with:
           bump_version_scheme: patch
           tag_prefix: v
-      - uses: actions/checkout@v2
+      - name: Checkout code
+        uses: actions/checkout@v2
       - name: Set up Python
         uses: actions/setup-python@v2
         with:
