### 2024-01-15 2e2233e
test: exception for optional torch typing import
