+++ dicaugment/augmentations/blur/functional.py
@@ -19,7 +19,6 @@ __all__ = [
     "blur",
     "median_blur",
     "gaussian_blur",
-    # "glass_blur"
 ]
 
 
@@ -31,6 +30,26 @@ def blur(
     mode: str = "constant",
     cval: Union[float, int] = 0,
 ) -> np.ndarray:
+    """Blur the input image using an mean kernel.
+
+    Args:
+        ksize (int): The kernel size for blurring the input image.
+        by_slice (bool): Whether the kernel should be applied by slice or the image as a whole. If true, a 2D kernel is convolved along each slice of the image.
+            Otherwise, a 3D kernel is used. Default: False
+        mode (str): scipy parameter to determine how the input image is extended during convolution to maintain image shape. Must be one of the following:
+
+            - `reflect` (d c b a | a b c d | d c b a): The input is extended by reflecting about the edge of the last pixel. This mode is also sometimes referred to as half-sample symmetric.
+            - `constant` (k k k k | a b c d | k k k k): The input is extended by filling all values beyond the edge with the same constant value, defined by the cval parameter.
+            - `nearest` (a a a a | a b c d | d d d d): The input is extended by replicating the last pixel.
+            - `mirror` (d c b | a b c d | c b a): The input is extended by reflecting about the center of the last pixel. This mode is also sometimes referred to as whole-sample symmetric.
+            - `wrap` (a b c d | a b c d | a b c d): The input is extended by wrapping around to the opposite edge.
+
+            Reference: https://docs.scipy.org/doc/scipy/reference/generated/scipy.ndimage.median_filter.html
+
+            Default: `constant`
+        cval (int,float): The fill value when mode = `constant`. Default: 0
+
+    """
     if by_slice:
         kernel = np.ones((ksize, ksize, 1), dtype=np.float32)
     else:
@@ -48,6 +67,26 @@ def median_blur(
     mode: str = "constant",
     cval: Union[float, int] = 0,
 ) -> np.ndarray:
+    """Blur the input image using median blue technique.
+
+    Args:
+        ksize (int): The kernel size for blurring the input image.
+        by_slice (bool): Whether the kernel should be applied by slice or the image as a whole. If true, a 2D kernel is convolved along each slice of the image.
+            Otherwise, a 3D kernel is used. Default: False
+        mode (str): scipy parameter to determine how the input image is extended during convolution to maintain image shape. Must be one of the following:
+
+            - `reflect` (d c b a | a b c d | d c b a): The input is extended by reflecting about the edge of the last pixel. This mode is also sometimes referred to as half-sample symmetric.
+            - `constant` (k k k k | a b c d | k k k k): The input is extended by filling all values beyond the edge with the same constant value, defined by the cval parameter.
+            - `nearest` (a a a a | a b c d | d d d d): The input is extended by replicating the last pixel.
+            - `mirror` (d c b | a b c d | c b a): The input is extended by reflecting about the center of the last pixel. This mode is also sometimes referred to as whole-sample symmetric.
+            - `wrap` (a b c d | a b c d | a b c d): The input is extended by wrapping around to the opposite edge.
+
+            Reference: https://docs.scipy.org/doc/scipy/reference/generated/scipy.ndimage.median_filter.html
+
[…]
