### 2024-01-23 b0e5509
chore: remove windows and macos testing
