### 2024-01-23 8015b35
docs: update documentation syntax
