+++ docs/source/installation.rst
@@ -13,7 +13,7 @@ Prerequisites
 
 Before installing DICaugment, ensure you have the following prerequisites:
 
-- Python 3.7 or higher
+- Python 3.8 or higher
 
 Installation Steps
 -------------------------
