### 2024-01-26 68142f0
Fix: Fix Failing Workflow (#8)
* chore: add automated testing and linter

* fix: add pytest-cov

* build(autotag): change master branch to main

* docs(docstring): add always_apply to docstrings

* docs(changelog): add changelog

* chore(template): add pull request template

* chore(auto_tag): change default bump to patch level

* chore(template): update pull request template

* chore(template): update typo in pull request template

* docs(paper): Update paper

* docs: update documentation (and other things)

* build: update workflow

* build: add dev branch

* build: fix workflow syntax

* te
[…]
