+++ .readthedocs.yml
@@ -15,6 +15,5 @@ build:
 
 
 python:
-  system_packages: true
   install:
     - requirements: docs/requirements.txt
