### 2024-01-28 a9a174b
Docs: Add Issue Templates (#10)
* docs(README): update README

* docs(templates): add issue templates

---------

Co-authored-by: jjmcintosh <jjmcintosh@github.com>
