### 2024-01-29 84c90b1
build: add manual workflow trigger
