+++ .github/workflows/auto_test.yml
@@ -3,6 +3,8 @@ name: automated testing
 on:
   pull_request:
     branches: [ "main" ]
+  
+  workflow_dispatch:
 
 permissions:
   contents: read
