### 2024-01-29 #6120 @jjmcintosh
@MaierOli2010 @kuadrat I have responded to your comments below and have made the necessary updates. Please let me know if anything is not satisfactory!

> I did not find any automated testing nor instructions on how to perform testing of the software. There is, however, a tests subfolder so I assume some testing is done in one way or another. Please either state how to perform the testing manually or, ideally, use some automated software to perform testing (like Jenkins or Travis).

I have added automated testing to this repo using GitHub Actions. You should be able to manually trigger the testing build in the actions tab in the repo by selecting the `automated testing` workflow and clicking on `Run workflow`

> There are no guidelines on how to report problems/bugs or how to contribute. Please include an appropriate document such as from this template: https://gist.github.com/briandk/3d2e8b3ec8daf5a27a62 and adapt it to your needs.

A CONTRIBUTING file has been added using that template along with issue templates.

> I've also noticed that not all parameters of the API are documented in each function. E.g. always_apply is not documented in all functions at https://dicaugment.readthedocs.io/en/latest/dicaugment.augmentations.html. Please make sure that all parameters of the public API are documented accordingly.

I have updated the documentation to address this.

> I do miss a broader overview of existing software for data augmentation. Albumentations is not the only package out there to do image augmentation. Some examples that I quickly found are TorchIO, MONAI, DLTK (also offers some augmentation techniques), MedicalTorch also claims to have some elastic transformations available to augment data. Finally, there is a tutorial from SimpleITK to apply data augmentation for 2D and 3D images available at https://simpleitk.org/SPIE2019_COURSE/03_data_augmentation.html. Please clearly state how you differentiate to these other tools, especially the claim of 3D augmentation seems to be covered at least by SimpleITK.
To this end, I also think that the list of references is not complete as these packages, and the corresponding publications if applicable, are certainly missing.

The paper has been updated with an overview of existing software for data augmentation and how Albumentations differs along with an updated set of references.

> While I can clearly see the contribution of the main author and also the last author as mentioned in the readme, the others are not to be found in any commit and are also not mentioned in your own readme. Can you maybe clarify the contributions? I do not doubt the contributions but it is hard to follow if there are no commits and even your own readme does not mention them as contributors.

Author contributions have been added in both the README and the paper.

> Testing the basic example in the readme gives an error:

Fixed. Thanks!

-------------------------------------------------------------------------------------------------------------------------------------
> Contribution and authorship @jjmcintosh's contributions are evident from the commit logs. Mehdi Farhangi is listed as an author in the README, though it is not evident from the repository nor paper, in which way both authors contributed. The same goes for the other authors listed in the paper. Please specify this, ideally in the README or a CONTRIBUTORS file and in the paper.

Author contributions have been added in both the README and the paper.

> Automated tests Running the tests after cloning the repository gave 206 failed, 4611 passed. Ensure that all tests pass in the version that is to be published. Also, the source code of the tests contains large numbers of commented out blocks of code, which indicates that some work on that front might still be required.

Thank you for opening a GitHub issue. I have addressed this issue. Also, the commented-out code blocks have been removed as they are irrelevant to this project.

> Documentation The tutorial part of the docs is clearly written, easy to follow and enhanced by well-made, relevant images. However, the API documentation contains a lot of undocumented functions and class method. Also, important methods such as the Composition class' __call__ method should be documented and it should be clear from the API doc alone - without the tutorial - that this is the intended way to make use of the class (if my understanding is correct here).
I can see that most Transform classes have methods of identical names, such as apply, which always serve the same purpose. Maybe [this stackoverflow answer](https://stackoverflow.com/a/65460426/5264510)) provides some possibilities for re-using documentation.

I have updated the missing docstrings. Each method in the documentation should have some text attached to it. I skewed away from adding arguments to the docstrings of the common methods (e.g `apply`) as to discourage the user from directly invoking that method and using the `Compose` pipeline. Please let me know what you think of this decision.

> State of the field I agree with the point brought up by @MaierOli2010 regarding the need to specify how DICaugment distinguishes itself from other 3D image augmentation software. I also found an additional package that seems to be quite similar: [volumentations](https://github.com/ZFTurbo/volumentations).

This has been updated in the paper!

> Typos
[this title](https://dicaugment.readthedocs.io/en/latest/bbox_augmentation.html#annotatation-formats) is misspelled
there's probably a newline missing in the docstring to [BBoxSafeRandomCrop](https://dicaugment.readthedocs.io/en/latest/augmentations.crops.html#dicaugment.augmentations.crops.transforms.BBoxSafeRandomCrop)
paper line 65: the units "mm" need not be italic
paper, caption of Fig. 1, second to last line: Siemens is misspelled

Fixed!

> Code coverage In order to assess whether the designed tests actually do a good job at testing the functionality of the code, I recommend the use of a test-coverage estimator, such as codecov

The GitHub actions testing pipeline has coverage!

-------------------------------------------------------------------------------------------------

### To-Do For Me

All of the non-required suggestions. I wanted to get the prerequisites out of the way before working on any optionals.
