### 2024-02-16 #6120 @osorensen
First, @jjmcintosh, could you please see if this suggested DOI is correct, and if so, add it to your BibTeX source?

```
MISSING DOIs

- 10.1118/1.4752209 may be a valid DOI for title: Quantitative comparison of noise texture across CT scanners
 from different manufacturers
```
