+++ paper/paper.md
@@ -38,7 +38,7 @@ DICaugment is a Python package based on the popular image augmentation library A
 
 # Statement of need
 
-The recent advancements in machine learning have significantly improved the performance of deep learning models across various domains and applications. However, the success of these models still largely rely on a significant amount of labeled and annotated training data. This is a particular limitation in medical imaging applications where imaging data annotation and the establishment of a reliable reference standard is expensive, time-consuming, and typically requires clinician expertise. DICaugment supports the efficient use of available labeled and annotated data by providing augmentations that are consistent with preexisting imaging data annotations available in the form of masks, bounding boxes, and keypoints. 
+The recent advancements in machine learning have significantly improved the performance of deep learning models across various domains and applications. However, the success of these models still largely relies on a significant amount of labeled and annotated training data. This is a particular limitation in medical imaging applications where imaging data annotation and the establishment of a reliable reference standard is expensive, time-consuming, and typically requires clinician expertise. DICaugment supports the efficient use of available labeled and annotated data by providing augmentations that are consistent with preexisting imaging data annotations available in the form of masks, bounding boxes, and keypoints. 
 
 Considering the distinct geometric and operational factors inherent to 3D images, several toolkits have been developed to perform transformations in the 3D space domain; SimpleITK [@Yaniv2017] provides a collection of spatial transformations, including linear and deformable geometric transformations, as well as intensity-based transformations such as linear, non-linear, and histogram-based intensity adjustments. In addition to supporting various medical imaging tasks, augmentation methods in packages such as TorchIO [@Pérez-García2021] and MONAI [@cardoso2022monai] were designed to seamlessly integrate into popular deep learning frameworks such as PyTorch, providing a convenient interface for integration within deep learning pipelines. Packages such as Volumentations [@solovyev20223d] were uniquely developed for the purpose of augmenting 3D volumetric images, offering specialized transformations for enriching 3D imaging datasets during the training of deep neural networks.
 
