### 2024-02-18 #6120 @jjmcintosh
@osorensen See [PR](https://github.com/DIDSR/DICaugment/pull/13) for adding that DOI. Just waiting on approval.
