### 2024-02-18 #6120 @jjmcintosh
> Thanks @jjmcintosh for carefully addressing the points made. For me, all the publication-critical issues are thus resolved and I would give the "go" for publication.
>
> I would still warmly encourage you to consider the _additional points_ I made above, as well as the below suggestions.
>
> # Non-critical
> * **Code coverage**: Great to see that code coverage is now automatically evaluated in a github workflow! Now that you have that, it would make sense to share the gist of the coverage report in some form in your repo. A common way to do this is by automatically sending the coverage report to [codecov](https://about.codecov.io/) and/or to generate a badge for your `README` that indicates the coverage results (see e.g. [here](https://stackoverflow.com/questions/59203704/how-to-create-a-badge-with-test-coverage-jacoco-on-github-actions), note also the second answer). This helps users and other developers spot at a glance that you have a well-working testing routine in place.
> * **Supported python version**: DICaugment supports python versions `3.8.0` and higher, which is explicitly stated in the installation instructions, reflected in your `setup.py` file and visible on [www.pipy.org](http://www.pipy.org). Unfortunately, this doesn't prevent users from "pip installing" DICaugment with unsupported python versions. To be even safer and avoid frustration on the user side, you could add something like:
>   ```
>   import sys
>
>   if sys.version_info < (3, 8):
>       raise Exception("DICaugment requires Python 3.8 or above.")
>   ```
>
>
>
>
>
>
>
>
>
>
>
>   to the `__init__.py` file and state the python version requirements in the `README` as well (potentially in the form of badges). I know this may feel quite redundant at this point, but I believe that it is not too much work and it would be well invested. In my experience, if a user encounters difficulties when downloading and installing software, chances are high that they are discouraged and driven away from ever using the package.
> * **Documentation**: It still seems to me that `Transoformation.apply()` (and similar functions) does not seem to be intended way for the package to be used. If this is correct, maybe the dilemma of unnecessarily documenting the same function over again could be avoided by making them private methods?

@kuadrat Working on your suggestions this weekend! The PyPi issue should be resolved now as I have deleted version `1.0.0`. You should be able to confirm this by installing in a fresh environment and using the `--no-cache-dir` flag
