### 2024-02-18 cbb79be
docs: Add DOI (#13)
* docs: add doi

---------

Co-authored-by: jjmcintosh <jjmcintosh@github.com>
