+++ paper/paper.bib
@@ -82,7 +82,8 @@
   month     =  oct,
   year      =  2012,
   copyright = "http://onlinelibrary.wiley.com/termsAndConditions",
-  language  = "en"
+  language  = "en",
+  doi       = {10.1118/1.4752209}
 }
 
 
