### 2024-02-19 #6120 @kuadrat
>
> @kuadrat Working on your suggestions this weekend! The PyPi issue should be resolved now as I have deleted version `1.0.0`. You should be able to confirm this by installing in a fresh environment and using the `--no-cache-dir` flag

I see, so an older version with different dependencies was causing this. I confirm that `pip install` for python 3.7.5 now throws an error, stating that `python>=3.8` is required.
