### 2024-02-25 #6120 @jjmcintosh
Hi all, still waiting on approvals from the GitHub Organization owners for the third-party OAuth apps (codecov and Zenodo). After those are approved, I will make a new release with those changes and post the version number and DOI here.
