+++ .github/workflows/auto_test.yml
@@ -38,4 +38,13 @@ jobs:
 
     - name: Test with pytest
       run: |
-        pytest --cov=.
+        pytest --cov=. --cov-report=xml
+
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v4.0.1
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        slug: DIDSR/DICaugment
+        file: ./coverage.xml
+        flags: unittests
+        verbose: true
