+++ .github/workflows/publish.yml
@@ -1,4 +1,4 @@
-name: Publish
+name: Build
 on:
   push:
     branches:
@@ -40,7 +40,16 @@ jobs:
 
     - name: Test with pytest
       run: |
-        pytest --cov=.
+        pytest --cov=. --cov-report=xml
+
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v4.0.1
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        slug: DIDSR/DICaugment
+        file: ./coverage.xml
+        flags: unittests
+        verbose: true
 
 
   create-release:
