### 2024-03-02 #6120 @jjmcintosh
@osorensen Updated! I misinterpreted the instructions and did not realize that the Zenodo archive had metadata attached to it that needed to be updated.
